bitkeeper revision 1.1236.58.1 (42527f89zgr9JJ0KbdZkIWCfo_KBfA)
SMP timer and irq fixes for 2.6. Merge x86/64 time.c with i386.

+++ linux-2.6.11-xen-sparse/arch/xen/i386/mm/hypervisor.c
@@ -108,7 +108,7 @@ void xen_tlb_flush_mask(cpumask_t mask)
 {
     struct mmuext_op op;
     op.cmd = MMUEXT_TLB_FLUSH_MULTI;
-    op.cpuset = (unsigned long)mask.bits;
+    op.cpuset = mask.bits;
     BUG_ON(HYPERVISOR_mmuext_op(&op, 1, NULL, DOMID_SELF) < 0);
 }
 
@@ -124,7 +124,7 @@ void xen_invlpg_mask(cpumask_t mask, unsigned long ptr)
 {
     struct mmuext_op op;
     op.cmd = MMUEXT_INVLPG_MULTI;
-    op.cpuset = (unsigned long)mask.bits;
+    op.cpuset = mask.bits;
     op.linear_addr = ptr & PAGE_MASK;
     BUG_ON(HYPERVISOR_mmuext_op(&op, 1, NULL, DOMID_SELF) < 0);
 }
